17-Sep-2016: Hi Robert
We have been asked to advertise this event on our website and in the surgery. I can see that it is in Basildon but the online survey address would be good to advertise as they can do that from home.
How easy is it to add to our Front Page? even an edited version.
Please can you let me know if that would be expensive to do.

I often get asked to advertise an event for patient feedback. How easy is it for you to display a temporary poster on our home page?

Thank you

Grainne