07-Dec-2016 email Hi Robert
These are our opening hours over Christmas and the New Year.

Please can these be put on the Home/front page. Please remove the GP Summit notice for 23/11/16. If you have to change the layout that is fine. The Patient Group notice could be in a text box and this new message in another box to separate them. Thank you

Kind regards

Grainne