From Grainne's email dated 01-Feb-2017
Completed the carousel part.
Dear Robert
I need to add the minutes of the meetings of the Patient Group for last
year to the website as I have not done so and a patient has complained.
I will send them in the next email.
Also I would like to add that we now have a Nurse Practitioner, Debbie
Guy who sees patients Monday and Friday. She can prescribe and has
experience working in the Minor Injury Unit for children and adults.
Kind regards
Grainne