See below
From Karen Allen on 17-Mar-2020:
Good Morning 
We need a notice on our website stating that:
Tollesbury Surgery has made the difficult decision of suspending open/walk  in surgery in the mornings, after advice from our peers and the CCG.

If you need an appointment and it is not related to coronavirus please phone the surgery and make an appointment. The GP will triage and either phone you or ask you to come to the surgery.

Thank you for your patience.​

Karen Allen
Practice Manager