See below
Email on 29-Apr-2020 (this was placing a link and image from Dr Link on to the carousel and in Online.html):
Hi Robert
Please read the information below and see attached.
Can you add this information onto our website please?
I think this is something like the start of what you saw earlier?
Regards
Karen Allen